refactor(db-structure): split monolithic Drizzle parser into modular components
Break down 1,336-line parser.ts into 9 focused modules for better maintainability:

- types.ts: Type definitions and interfaces
- ast-utils.ts: AST manipulation utilities
- expression-parser.ts: Expression parsing logic
- column-parser.ts: Column definition parsing
- table-parser.ts: Table structure parsing
- enum-parser.ts: Enum definition parsing
- relation-parser.ts: Relation definition parsing
- converter.ts: Data conversion logic
- main-parser.ts: Main orchestrator

Each module now has a single responsibility, improving testability and readability.
